Fixed unit-tests script
Summary: Fixed unit-tests script

Test Plan: test

Reviewers: ali

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D801

+++ unit-tests/thread_basic.c
@@ -7,9 +7,9 @@
 void *
 routine(void *arg)
 {
-    int thrNo = (int)arg;
+    intptr_t thrNo = (intptr_t)arg;
 
-    printf("Hello %d\n", thrNo);
+    printf("Hello %ld\n", thrNo);
 
     return NULL;
 }
